backend/requirements.txt
flask==3.1.0
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
pyjwt==2.10.1
